Display the turtle output in an SVG on the page for pyodide
This took way longer than it should have to figure out. I asked a
StackOverflow question about this here:

https://stackoverflow.com/questions/77753482/how-to-render-to-an-animated-svg-on-the-page-using-pyodide-and-the-basthon-turtl/77753689

Then eventually figured out how to display the output based on this
web worker example:

https://pyscript.com/@ntoll/turtle-workers/latest

For some reason, the turtle is displaying as an icon of a turtle in the
pyodide implementation and as just an arrow in the skulpt implementation
but that’s probably configurable. I’m a bit concerned that this pyodide
turtle package might not support all the drawing functions we need since
there are a lot of methods that aren’t implemented:

https://framagit.org/basthon/basthon-kernel/-/blob/master/packages/kernel-python3/src/modules/turtle/turtle/__init__.py#L540-616

Although, perhaps those aren’t important because they’re methods on
instances of the Screen() object.