Build all of the basthon project’s packages (not just turtle)
We also need to support p5 which is fortunately one of the packages that
has been adapted to work with pyodide by the Basthon project. Therefore,
Replace the ./bin/build_turtle script with a ./bin/build_packages script
that builds all of the available packages that Basthon supports. These
might not all be needed and some might require more work to get the
output to display on the page, but we can investigate this later.

https://framagit.org/basthon/basthon-kernel/-/tree/master/packages/kernel-python3/src/modules