Ensure the Makefile is called for each Basthon package
Previously, we were calling python3 -m build for each package, but some
of the packages have additional steps. For example, the p5 package
uses gulp to minify the p5.js file to p5.min.js file which is then
included in the wheel. Otherwise, the wheel fails to import in pyodide.

https://framagit.org/basthon/basthon-kernel/-/blob/master/packages/kernel-python3/src/modules/p5/Makefile