Allow the p5 package to be imported
The p5 package from Basthon depends on a ‘basthon’ package. I’m not sure
if any of this is needed yet, so provide a fake implementation of that
package in the pyodide environment.

This GitHub comment led me in a useful direction:
https://github.com/berinhard/pyp5js/issues/124#issuecomment-733325510

kernel.display_event() is called from this file:
https://framagit.org/basthon/basthon-kernel/-/blob/master/packages/kernel-python3/src/modules/p5/p5/_core.py#L82-83

kernel.locals().get() and kernel.locals().keys() are called from this file:
https://framagit.org/basthon/basthon-kernel/-/blob/master/packages/kernel-python3/src/modules/p5/p5/_nsmanager.py#L25