Make the basthon package support rendering to the web page
In order to allow p5 to render to the web page, we need to provide
slightly more realistic implementations for locals and display_event:

- locals: This function needs to provide a list of all globals in the
top-level namespace. Otherwise, the p5 library can’t find the setup and
draw functions, which is part of its standard interface. See here:

https://framagit.org/basthon/basthon-kernel/-/blob/master/packages/kernel-python3/src/basthon/kernel.py#L43-45

- display_event: This function is the mechanism that Basthon packages
use to signal that they want to display a HTML node on the page. In our
case, we should replace the contents of the #visual div with the node.

https://framagit.org/basthon/basthon-kernel/-/blob/master/packages/kernel-python3/src/basthon/kernel.py#L48-59