Use the same display_event method for the turtle package
Previously, we were using the js package from within python to set
visual.innerHTML but it’s probably better to use the same standardised
interface that the p5 package uses (display_event) for signalling the
element that should be displayed. This might help later if we need to
move pyodide execution into a web worker to support interruptions and
for better isolation from the outer web page.

The Basthon project does actually patch the turtle package to make
turtle.show_scene() call the display_event function, but we’re not
applying these patches because we’re using a fake implementation of the
basthon package. I don’t think we need any of the other patch code and
it would be fairly easy to implement later if we did, e.g. download.

https://framagit.org/basthon/basthon-kernel/-/blob/master/packages/kernel-python3/src/basthon/_patch_modules.py#L128-148

https://pyodide.org/en/stable/usage/keyboard-interrupts.html