Load packages from PyPi if they are not vendored
According to the Pyodide docs, it won’t automatically import packages
from PyPi although this is supported using micropip. Therefore, add a
loadDependency function that can also load packages from PyPi.

Some of the imports will be for modules built into Python or for
packages that come with Pyodide so we should check this first.

We still need to call loadPackagesFromImports because Pyodide doesn’t
seem to load sqlite3 when loadPackage is called. I think this is because
sqlite3 isn’t included by default, but that doesn’t explain why
loadPackagesFromImports is able to load it whereas loadPackage isn’t.

https://pyodide.org/en/stable/usage/loading-packages.html

https://pyodide.org/en/stable/usage/wasm-constraints.html