Fix sense_hat/__init__.py to import PIL instead of image
It looks like the image module was removed from Python and replaced with
a separate pillow package that provides a PIL import. Either that, or
Skulpt provided an image module that is not available in Python. I can
see that that the interface for PIL looks correct as it contains a
getData function that is called elsewhere in __init__.py

https://omz-software.com/pythonista/docs/ios/Image.html