Stop the program if Stop was pressed while packages are loading
When runPython is called, it seems to reset the interruptBuffer to 0. If
the user pressed the Stop button before runPython is called while
packages are loading, it would still call runPython and the program
wouldn’t be stopped.

Therefore, introduce a separate ‘stopPython’ method that explicitly sets
a stopped state variable and then check this repeatedly while packages
are loading. Throw a KeybordInterrupt so that the error matches the type
thrown by pyodide when the program is interrupted.