Emit ‘handleError’ messages if the Python program errors
If something else errors in JavaScript, re-throw the error. If the
program is interrupted during package loading, throw a PythonError so
that it is handled in the same way as interruptions that occur when the
program is running.

Present the error in the UI under all other output. If the error type
is KeyboardInterrupt then the error won’t contain a line number or
content so show “Program interrupted” instead.