Add support for capturing standard input in the Pyodide prototype
This is pretty complicated. Firstly, the pyodide.setStdin(…) function
only allows a synchronous handler which means we need to synchronously
block inside the web worker until input is received from the web page.
That means, we can’t call postMessage back from the web page because the
onmessage function isn’t re-evaluated until the synchronous call has
ended, which is too late.

Instead, we have to send standard input back from the web page using a
SharedArrayBuffer and use Atomics.wait to poll an index within this
buffer to check if it has changed. Additionally, we are unable to grow
the SharedArrayBuffer in Firefox so need to allocate enough memory for
the contents of the buffer in advance. 1 MiB should be more than enough.

The simplest way I could think to implement this is to store the length
of actual content within the array in index 0 of the SharedArrayBuffer
and then check this index for changes in Atomics.wait. When we
eventually receive input from the web page, it is responsible for
updating the length of the array into index 0 so that the web worker can
see the change and continue.

https://pyodide.org/en/stable/usage/api/js-api.html#pyodide.setStdin